visual_service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
torch
ultralytics
transformers
pillow
